Convert controller text to paths; fixes #20
The controller svg uses flow text for the labels. Browsers can't handle this though. Thus, I converted the flow text to paths. Even though there is also the text which is supported by browsers, using paths provides the advantage that the client does not need to have the used font installed.